Enter a new note

+++ Dorset_OOP_Project/Note.cs
@@ -7,8 +7,8 @@ namespace Dorset_OOP_Project
     public class Note
     {
         public Exam ExamNote { get; set; }
-        public float NoteValue { get; set; }
-        public Note(Exam _examNote,float _noteValue)
+        public double NoteValue { get; set; }
+        public Note(Exam _examNote,double _noteValue)
         {
             ExamNote = _examNote;
             NoteValue=_noteValue;
